+++ NEWS.md
@@ -5,6 +5,8 @@
 * add a `null_action` argument to `val_labels()`, `val_label()` and a 
   `.null_action` argument to `set_value_labels()`, `add_value_labels()` and
   `remove_value_labels()` (#145)
+* new functions `update_variable_labels_with()` and `update_value_labels_with()`
+  allowing to update variable/value labels with a custom function (#153)
 
 **Bug fix**
 
@@ -12,7 +14,7 @@
   too small (#148)
 * fix in `recode.haven_labelled()` when `.x` contains `NA` and
   `.combine_value_labels = TRUE` (#151)
-* producing an error when trying to assign value labels to a Date vector (#156)
+* produce an error when trying to assign value labels to a Date vector (#156)
 
 # labelled 2.12.0